+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Yii View Change Log
 
+## 4.1.0 under development
 
-## 4.0.1 under development
-
-- no changes in this release.
+- New #193: Add simple view context class `ViewContext` (vjik)
+- New #193: Add methods `View::withContextPath()` and `WebView::withContextPath()` that set view context path (vjik)
 
 ## 4.0.0 October 25, 2021